requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
langchain-core==0.3.15
langchain-groq==0.2.1
neo4j==5.27.0
requests==2.31.0
numpy==1.26.4
scipy==1.11.4
